requirements.txt
fastapi[standard]
uvicorn
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
pydantic==2.11.7
pydantic-settings
python-dotenv
python-multipart==0.0.6
opencv-python
pillow==10.4.0
marker-pdf==1.7.5
torch
torchvision
pandas==2.1.4
openpyxl
python-dateutil
regex
